Make some Api public
