+++ src/Microsoft.Restier.Publishers.OData/Model/EdmHelpers.cs
@@ -4,15 +4,105 @@
 using System;
 using System.Globalization;
 using System.Linq;
+using System.Web.OData;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.OData.Edm;
 
 namespace Microsoft.Restier.Publishers.OData.Model
 {
-    internal static class EdmHelpers
+    /// <summary>
+    /// This class contains some common extension methods for Edm
+    /// </summary>
+    public static class EdmHelpers
     {
         private const string DefaultEntityContainerName = "DefaultContainer";
 
-        public static EdmEntityContainer EnsureEntityContainer(this EdmModel model, Type apiType)
+        /// <summary>
+        /// The type to get the primitive type reference
+        /// </summary>
+        /// <param name="type">The clr type to get edm type reference</param>
+        /// <returns>The edm type reference for the clr type</returns>
+        public static EdmTypeReference GetPrimitiveTypeReference(this Type type)
+        {
+            // Only handle primitive type right now
+            bool isNullable;
+            EdmPrimitiveTypeKind? primitiveTypeKind = EdmHelpers.GetPrimitiveTypeKind(type, out isNullable);
+
+            if (!primitiveTypeKind.HasValue)
+            {
+                return null;
+            }
+
+            return new EdmPrimitiveTypeReference(
+                EdmCoreModel.Instance.GetPrimitiveType(primitiveTypeKind.Value),
+                isNullable);
+        }
+
+        /// <summary>
+        /// Get the clr type for a specified edm type
+        /// </summary>
+        /// <param name="edmType">The edm type to get clr type</param>
+        /// <param name="serviceProvider">The provider to get service from DI container</param>
+        /// <returns>The clr type</returns>
+        public static Type GetClrType(this IEdmType edmType, IServiceProvider serviceProvider)
+        {
+            IEdmModel edmModel = serviceProvider.GetService<IEdmModel>();
+
+            ClrTypeAnnotation annotation = edmModel.GetAnnotationValue<ClrTypeAnnotation>(edmType);
+            if (annotation != null)
+            {
+                return annotation.ClrType;
+            }
+
+            throw new NotSupportedException(string.Format(
+                CultureInfo.InvariantCulture,
+                Resources.ElementTypeNotFound,
+                edmType.FullTypeName()));
+        }
+
+        /// <summary>
+        /// Get the edm type reference for a clr type
+        /// </summary>
+        /// <param name="type">The clr type</param>
+        /// <param name="model">The Edm model</param>
+        /// <returns>The Edm type reference</returns>
+        public static IEdmTypeReference GetTypeReference(this Type type, IEdmModel model)
+        {
+            if (type == null || model == null)
+            {
+                return null;
+            }
+
+            Type elementType;
+            if (type.TryGetElementType(out elementType))
+            {
+                return EdmCoreModel.GetCollection(GetTypeReference(elementType, model));
+            }
+
+            var edmType = model.FindDeclaredType(type.FullName);
+
+            var enumType = edmType as IEdmEnumType;
+            if (enumType != null)
+            {
+                return new EdmEnumTypeReference(enumType, true);
+            }
+
+            var complexType = edmType as IEdmComplexType;
+            if (complexType != null)
+            {
+                return new EdmComplexTypeReference(complexType, true);
+            }
+
+            var entityType = edmType as IEdmEntityType;
+            if (entityType != null)
+            {
+                return new EdmEntityTypeReference(entityType, true);
+            }
+
+            return type.GetPrimitiveTypeReference();
+        }
+
+        internal static EdmEntityContainer EnsureEntityContainer(this EdmModel model, Type apiType)
         {
             var container = (EdmEntityContainer)model.EntityContainer;
             if (container == null)
@@ -24,7 +114,7 @@ namespace Microsoft.Restier.Publishers.OData.Model
             return container;
         }
 
-        public static IEdmEntitySet FindDeclaredEntitySetByTypeReference(
+        internal static IEdmEntitySet FindDeclaredEntitySetByTypeReference(
             this IEdmModel model, IEdmTypeReference typeReference)
         {
             IEdmTypeReference elementTypeReference;
@@ -40,22 +130,6 @@ namespace Microsoft.Restier.Publishers.OData.Model
 
             return model.EntityContainer.EntitySets()
                 .SingleOrDefault(e => e.EntityType().FullTypeName() == elementTypeReference.FullName());
-        }
-
-        public static EdmTypeReference GetPrimitiveTypeReference(this Type type)
-        {
-            // Only handle primitive type right now
-            bool isNullable;
-            EdmPrimitiveTypeKind? primitiveTypeKind = EdmHelpers.GetPrimitiveTypeKind(type, out isNullable);
-
-            if (!primitiveTypeKind.HasValue)
-            {
-                return null;
-            }
-
-            return new EdmPrimitiveTypeReference(
-                EdmCoreModel.Instance.GetPrimitiveType(primitiveTypeKind.Value),
-                isNullable);
         }
 
         private static bool TryGetElementTypeReference(
